ADD notification ADD List of followers & followed FIX button follow user

+++ app/Http/Controllers/UserFollowController.php
@@ -12,6 +12,8 @@ use Illuminate\Support\Facades\Redirect;
 use Illuminate\Validation\Rules;
 use Illuminate\View\View;
 
+use App\Notifications\RealTimeNotification;
+
 use App\Models\Follow;
 use App\Models\Tweet;
 use App\Models\User;
@@ -25,7 +27,12 @@ class UserFollowController extends Controller
      */
     public function dashboardFollowed(): View
     {
-        $tweets = Tweet::whereIn('user_id', Auth::guard('user')->user()->following->pluck('id'))->latest()->with(['user', 'likes'])->paginate(10); // ->dumpRawSql();
+        $tweets = Tweet::whereIn('user_id', Auth::guard('user')->user()->followers->pluck('followed_user_id'))->latest()->with(['user', 'likes'])->paginate(10); // ->dumpRawSql();
+
+        foreach ($tweets as $tweet)
+        {
+            $tweet->tweet = $this->hashtag_links($tweet->tweet);
+        }
 
         return $this->renderView('tweet.followed', [ 'tweets' => $tweets ]);
     }
@@ -36,14 +43,21 @@ class UserFollowController extends Controller
             return response('You cannot follow yourself...', 422);
         }
 
-        if (Auth::guard('user')->user()->following->contains($user->id)) {
+        if (Auth::guard('user')->user()->followers->pluck('follower_user_id')->contains($user->id)) {
             return response('You cannot follow more than one time', 409);
         }
 
         Follow::create([
-            "follower_user_id" => Auth::guard('user')->user()->id,
-            "followed_user_id" => $user->id,
-        ]); 
+            "follower_user_id" => $user->id,
+            "followed_user_id" => Auth::guard('user')->user()->id,
+        ]);
+
+        $user->notify(
+            new RealTimeNotification(
+                '@'.Auth::guard('user')->user()->name.' vous a follow !',
+                route('user.profile', ['user' => Auth::guard('user')->user()->name])
+            )
+        );
 
         return back();
     }
@@ -51,11 +65,23 @@ class UserFollowController extends Controller
     public function destroy(User $user)
     {
         $follow = Follow::query()
-                ->where('follower_user_id', auth()->id())
-                ->where('followed_user_id', $user->id)
+                ->where('followed_user_id', auth()->id())
+                ->where('follower_user_id', $user->id)
                 ->first();
 
+        if(!$follow)
+        {
+            abort(403);
+        }
+
         $follow->delete();
+
+        $user->notify(
+            new RealTimeNotification(
+                '@'.Auth::guard('user')->user()->name.' vous a unfollow !',
+                route('user.profile', ['user' => Auth::guard('user')->user()->name])
+            )
+        );
 
         return back();
     }
